Improve Stat struct

+++ app/stat.go
@@ -7,11 +7,11 @@ import (
 
 // Stat is a JSON-serializable struct for the /stat endpoint.
 type Stat struct {
-	FileSize                int64  `json:"fileSize"`
-	CurrentReportEventCount uint32 `json:"currentReportEventCount"`
-	LastReportEventCount    uint32 `json:"lastReportEventCount"`
-	LastReportDuration      string `json:"lastReportDuration"`
-	LastReportTime          int32  `json:"lastReportTime"`
+	FileSize                int64  `json:"fileSize"`                // in bytes
+	CurrentReportEventCount uint32 `json:"currentReportEventCount"` // number of events in the current report so far
+	LastReportEventCount    uint32 `json:"lastReportEventCount"`    // number of events in the last report
+	LastReportDuration      string `json:"lastReportDuration"`      // duration of the last report
+	LastReportTime          int64  `json:"lastReportTime"`          // Unix timestamp of the last report
 }
 
 // handleGetStat is the HTTP handler for the /stat endpoint.
@@ -22,7 +22,7 @@ func (a *App) handleGetStat(w http.ResponseWriter, r *http.Request) {
 		CurrentReportEventCount: a.reportRunner.CurrentReportEventCount(),
 		LastReportEventCount:    a.reportRunner.LastReportEventCount(),
 		LastReportDuration:      a.reportRunner.LastReportDuration().String(),
-		LastReportTime:          int32(a.reportRunner.LastReportTime().Unix()),
+		LastReportTime:          a.reportRunner.LastReportTime().Unix(),
 	}
 	data, err := json.MarshalIndent(s, "", "  ")
 	if err != nil {
